remove unused types
