+++ packages/insomnia/src/network/grpc/index.ts
@@ -2,6 +2,7 @@ import * as grpc from '@grpc/grpc-js';
 import { Call, ServiceError } from '@grpc/grpc-js';
 import { ServiceClient } from '@grpc/grpc-js/build/src/make-client';
 
+import { ResponseCallbacks } from '../../main/ipc/grpc';
 import * as models from '../../models';
 import type { GrpcRequest, GrpcRequestHeader } from '../../models/grpc-request';
 import { SegmentEvent, trackSegmentEvent } from '../../ui/analytics';
@@ -11,25 +12,6 @@ import { getMethodType, GrpcMethodTypeEnum } from './method';
 import parseGrpcUrl from './parse-grpc-url';
 import type { GrpcIpcMessageParams, GrpcIpcRequestParams } from './prepare';
 import * as protoLoader from './proto-loader';
-import { ResponseCallbacks } from './response-callbacks';
-
-const _createClient = (
-  req: GrpcRequest,
-  respond: ResponseCallbacks,
-): ServiceClient | undefined => {
-  const { url, enableTls } = parseGrpcUrl(req.url);
-
-  if (!url) {
-    respond.sendError(req._id, new Error('URL not specified'));
-    return undefined;
-  }
-
-  const credentials = enableTls ? grpc.credentials.createSsl() : grpc.credentials.createInsecure();
-  console.log(`[gRPC] connecting to url=${url} ${enableTls ? 'with' : 'without'} TLS`);
-  // @ts-expect-error -- TSCONVERSION second argument should be provided, send an empty string? Needs testing
-  const Client = grpc.makeGenericClientConstructor({});
-  return new Client(url, credentials);
-};
 
 const _makeUnaryRequest = (
   {
@@ -43,20 +25,15 @@ const _makeUnaryRequest = (
 ): Call | undefined => {
   // Create callback
   const callback = _createUnaryCallback(requestId, respond);
-
   // Load initial message
   const messageBody = _parseMessage(bodyText, requestId, respond);
-
   if (!messageBody) {
     return;
   }
-
   const grpcMetadata = _parseMetadata(metadata, requestId, respond);
-
   if (!grpcMetadata) {
     return;
   }
-
   // eslint-disable-next-line consistent-return
   return client.makeUnaryRequest(
     path,
@@ -77,13 +54,10 @@ const _makeClientStreamRequest = ({
 }: RequestData): Call | undefined => {
   // Create callback
   const callback = _createUnaryCallback(requestId, respond);
-
   const grpcMetadata = _parseMetadata(metadata, requestId, respond);
-
   if (!grpcMetadata) {
     return;
   }
-
   // Make call
   return client.makeClientStreamRequest(path, requestSerialize, responseDeserialize, grpcMetadata, callback);
 };
@@ -100,17 +74,13 @@ const _makeServerStreamRequest = (
 ): Call | undefined => {
   // Load initial message
   const messageBody = _parseMessage(bodyText, requestId, respond);
-
   if (!messageBody) {
     return;
   }
-
   const grpcMetadata = _parseMetadata(metadata, requestId, respond);
-
   if (!grpcMetadata) {
     return;
   }
-
   // Make call
   const call = client.makeServerStreamRequest(
     path,
@@ -121,7 +91,6 @@ const _makeServerStreamRequest = (
   );
 
   _setupServerStreamListeners(call, requestId, respond);
-
   // eslint-disable-next-line consistent-return
   return call;
 };
@@ -134,16 +103,12 @@ const _makeBidiStreamRequest = ({
   metadata,
 }: RequestData): Call | undefined => {
   const grpcMetadata = _parseMetadata(metadata, requestId, respond);
-
   if (!grpcMetadata) {
     return;
   }
-
   // Make call
   const call = client.makeBidiStreamRequest(path, requestSerialize, responseDeserialize, grpcMetadata);
-
   _setupServerStreamListeners(call, requestId, respond);
-
   return call;
 };
 
@@ -154,7 +119,21 @@ interface RequestData {
   method: GrpcMethodDefinition;
   metadata: GrpcRequestHeader[];
 }
-
+const _createClient = (
+  req: GrpcRequest,
+  respond: ResponseCallbacks,
+): ServiceClient | undefined => {
+  const { url, enableTls } = parseGrpcUrl(req.url);
+  if (!url) {
+    respond.sendError(req._id, new Error('URL not specified'));
+    return undefined;
+  }
+  const credentials = enableTls ? grpc.credentials.createSsl() : grpc.credentials.createInsecure();
+  console.log(`[gRPC] connecting to url=${url} ${enableTls ? 'with' : 'without'} TLS`);
+  // @ts-expect-error -- TSCONVERSION second argument should be provided, send an empty string? Needs testing
+  const Client = grpc.makeGenericClientConstructor({});
+  return new Client(url, credentials);
+};
 export const start = async (
   { request }: GrpcIpcRequestParams,
   respond: ResponseCallbacks,
@@ -162,7 +141,6 @@ export const start = async (
   const requestId = request._id;
   const metadata = request.metadata;
   const method = await protoLoader.getSelectedMethod(request);
-
   if (!method) {
     respond.sendError(
       requestId,
@@ -170,16 +148,12 @@ export const start = async (
     );
     return;
   }
-
   const methodType = getMethodType(method);
-
   // Create client
   const client = _createClient(request, respond);
-
   if (!client) {
     return;
   }
-
   const requestParams: RequestData = {
     requestId,
     client,
@@ -188,53 +162,44 @@ export const start = async (
     metadata,
   };
   let call;
-
   switch (methodType) {
     case GrpcMethodTypeEnum.unary:
       call = _makeUnaryRequest(requestParams, request.body.text || '');
       break;
-
     case GrpcMethodTypeEnum.server:
       call = _makeServerStreamRequest(requestParams, request.body.text || '');
       break;
-
     case GrpcMethodTypeEnum.client:
       call = _makeClientStreamRequest(requestParams);
       break;
-
     case GrpcMethodTypeEnum.bidi:
       call = _makeBidiStreamRequest(requestParams);
       break;
-
     default:
       return;
   }
-
   if (!call) {
     return;
   }
-
   // Update request stats
   models.stats.incrementExecutedRequests();
   trackSegmentEvent(SegmentEvent.requestExecute);
-
   _setupStatusListener(call, requestId, respond);
-
   respond.sendStart(requestId);
   // Save call
   callCache.set(requestId, call);
 };
-
+const _setupStatusListener = (call: Call, requestId: string, respond: ResponseCallbacks) => {
+  call.on('status', s => respond.sendStatus(requestId, s));
+};
 export const sendMessage = (
   { body, requestId }: GrpcIpcMessageParams,
   respond: ResponseCallbacks,
 ) => {
   const messageBody = _parseMessage(body.text || '', requestId, respond);
-
   if (!messageBody) {
     return;
   }
-
   // HACK BUT DO NOT REMOVE
   // this must happen in the next tick otherwise the stream does not flush correctly
   // Try removing it and using a bidi RPC and notice messages don't send consistently
@@ -249,16 +214,11 @@ export const commit = (requestId: string) => callCache.get(requestId)?.end();
 export const cancel = (requestId: string) => callCache.get(requestId)?.cancel();
 export const cancelMultiple = (requestIds: string[]) => requestIds.forEach(cancel);
 
-const _setupStatusListener = (call: Call, requestId: string, respond: ResponseCallbacks) => {
-  call.on('status', s => respond.sendStatus(requestId, s));
-};
-
 const _setupServerStreamListeners = (call: Call, requestId: string, respond: ResponseCallbacks) => {
   call.on('data', data => respond.sendData(requestId, data));
   call.on('error', (error: ServiceError) => {
     if (error && error.code !== grpc.status.CANCELLED) {
       respond.sendError(requestId, error);
-
       // Taken through inspiration from other implementation, needs validation
       if (error.code === grpc.status.UNKNOWN || error.code === grpc.status.UNAVAILABLE) {
         respond.sendEnd(requestId);
@@ -288,7 +248,6 @@ const _createUnaryCallback = (requestId: string, respond: ResponseCallbacks) => 
     // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
     respond.sendData(requestId, value!);
   }
-
   respond.sendEnd(requestId);
   callCache.clear(requestId);
 };
